pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "quickcan"
version = "0.1.0"
description = "QuickCAN - USB to CAN Python interface"
authors = [
  { name="Alok Mishra", email="alokm014@gmail.com" }
]
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.7"

dependencies = [
  "pyserial",
  "python-can"
]

[project.scripts]
quickcan-cli = "cli.quickcan_cli:main"

[tool.setuptools.packages.find]
where = ["."]
